Restores `ProofRequest.image_hash`, which #4321 removed on the grounds that
it had "no production reader". This branch adds that reader: the challenger
sets it from the disputed game's own `TEE_IMAGE_HASH`, and `NitroEnclavePool`
selects a registered enclave whose on-chain `signerImageHash` matches, failing
the job when none does. Without it, a host holding enclaves of more than one
image spends a full proof run before the mismatch surfaces as an on-chain
revert from `TEEVerifier.verify`.

The rest of #4321 stands. The proposer keeps none of its deleted image-hash
plumbing: it creates games at the current image, so it sends `B256::ZERO`,
which `select_enclaves_for_image` now reads as "no expectation" and satisfies
with any registered enclave. Only the challenger, which disputes games created
under possibly-retired images, names one.

+++ crates/execution/eip8130/src/tx_error.rs
@@ -31,7 +31,7 @@ pub enum TxAuthError {
         /// The operation whose scope requirement was not met.
         operation: Operation,
         /// The resolved actor's scope bitfield.
-        scope: u8,
+        scope: u16,
     },
 
     /// A config change or delegation targets a locked account. Both operations
@@ -40,9 +40,9 @@ pub enum TxAuthError {
     #[error("account is locked")]
     AccountLocked,
 
-    /// A delegation was not authorized by the account's native secp256k1 admin
-    /// self actor.
-    #[error("delegation requires the native-k1 admin self actor")]
+    /// A delegation was not authorized by an admin (unrestricted) actor on the
+    /// unlocked account.
+    #[error("delegation requires an admin actor")]
     DelegationUnauthorized,
 
     /// A config change is bound to a chain other than `0` (multichain) or the
